third_party/WebKit/LayoutTests/platform/win/fast/events/wheel/wheel-scroll-latching-on-scrollbar-expected.txt

CONSOLE ERROR: line 2487: Uncaught Error: assert_equals: The parent div is not supposed to scroll. expected 0 but got 33
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: The parent div is not supposed to scroll. expected 0 but got 33
PASS Scrolling on the child div scrollbar doesn't propagate to the parent div 
Harness: the test ran to completion.
